% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmre.R
\name{osmre_context}
\alias{osmre_context}
\title{Return context: full source of this client}
\usage{
osmre_context()
}
\value{
character: source code of this file
}
\description{
Return context: full source of this client
}
