% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmre.R
\name{osmre_offices}
\alias{osmre_offices}
\title{Fetch OSMRE office locations}
\usage{
osmre_offices()
}
\value{
tibble: office_name, address, city, state, zip, osm_region, web_site, lat, lon
}
\description{
Fetch OSMRE office locations
}
